build(auditlog): auditlog removal and moving components to shared
