--- libs/auditlog/src/index.ts
@@ -1,1 +1,0 @@
-export { AuditLogPage } from "./pages/AuditLogPageTemp";
