--- libs/auditlog/src/logs/index.ts
@@ -1,2 +1,0 @@
-export * from "./AuditLogsDataTable";
-export * from "./AuditLogsTimeline";
