--- libs/auditlog/tsconfig.json
@@ -1,34 +1,0 @@
-{
-	"compilerOptions": {
-		/* Lang and Env */
-		"lib": ["ES2022", "dom"],
-		"target": "es6",
-		"jsx": "react-jsx",
-		// "types": ["vitest/globals", "@testing-library/jest-dom/vitest"],
-
-		/* Modules */
-		"module": "esnext",
-		"moduleResolution": "bundler",
-		"allowSyntheticDefaultImports": true,
-		"resolveJsonModule": true,
-
-		/* Emit */
-		"declaration": true,
-		"outDir": "./dist",
-		"declarationDir": "./dist/types",
-		"emitDeclarationOnly": true,
-
-		/* Skip */
-		"skipLibCheck": true,
-		"allowJs": true,
-		"esModuleInterop": true,
-		"isolatedModules": true,
-
-		/* Lang and Env */
-		"paths": {
-			"@/*": ["./src/*"]
-		}
-	},
-	"include": ["src"],
-	"exclude": ["dist", "node_modules", "**/*.spec.ts*"]
-}
