+++ libs/shared/package.json
@@ -11,11 +11,17 @@
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "flexlayout-react": "^0.8.1",
-    "lucide-react": "^0.545.0"
+    "lucide-react": "^0.545.0",
+    "echarts": "^5.5.1",
+    "echarts-for-react": "^3.0.2",
+    "@mui/icons-material": "6.0.0"
   },
   "devDependencies": {
     "@types/react": "18.3.1",
     "@types/react-dom": "18.3.0",
     "typescript": "~5.5.2"
+  },
+  "peerDependencies": {
+    "@mui/icons-material": "6.0.0"
   }
 }
