+++ libs/shared/src/components/auditlog/index.ts
@@ -1,0 +1,4 @@
+import { AuditLogsDataTable } from "./logs/AuditLogsDataTable";
+import { AuditLogsTimeline } from "./logs/AuditLogsTimeline";
+
+export { AuditLogsDataTable, AuditLogsTimeline };
