+++ libs/shared/src/components/auditlog/logs/AuditLogsTimeline.tsx
@@ -13,8 +13,8 @@ import {
 	styled,
 	Typography,
 } from "@semoss/ui";
-import type { EventData } from "../utility/common";
-import { TimeDateFormatter } from "../utility/common";
+import type { EventData } from "../common";
+import { TimeDateFormatter } from "../common";
 
 // Type definitions
 interface ProcessedEventData {
