+++ libs/shared/src/components/index.ts
@@ -1,3 +1,4 @@
+export * from "./auditlog";
 export * from "./engine";
 export * from "./flex-layout";
 export * from "./form";
